% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicRegion_OccurrenceTable.R
\name{anatomicRegion}
\alias{anatomicRegion}
\title{'anatomicRegion'}
\usage{
anatomicRegion(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-anatomicRegion(DICOMFolderPath)}
}
\value{
A dataframe representing anatomic region of each shoot
}
\description{
anatomicRegion function represents anatomic region of each shoot
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
anatomicRegion(DICOMList)
}
