% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_OccurrenceTable.R
\name{device}
\alias{device}
\title{'device'}
\usage{
device(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating manufacturer of device which took a shoot
}
\description{
device function indicates manufacturer of device which took a shoot
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
device(DICOMList)
}
