% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICOMHeaderList.R
\name{DICOMHeaderList}
\alias{DICOMHeaderList}
\title{'DICOMHeaderList'}
\usage{
DICOMHeaderList(DICOMFolderPath, core = 4)
}
\arguments{
\item{DICOMFolderPath}{function will read all of the DICOM files under this pathway}

\item{core}{you can set CPU core in order to make the computing process fast}
}
\value{
A list containing metadata of DICOM
}
\description{
DICOMHeaderList function will read metadata of all DICOM files under DICOMFolderPath
}
\examples{
DICOMHeaderList(DICOMFolderPath, core=4)
}
