% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageNo_ImageTable.R
\name{imageNo}
\alias{imageNo}
\title{'imageNo'}
\usage{
imageNo(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list indicating order of DICOM images in specific radiologyOccurrenceId, and radiologyPhaseConceptId
}
\description{
imageNo function will indicate order of DICOM images in specific radiologyOccurrenceId, and radiologyPhaseConceptId
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageNo(DICOMList)
}
