% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageOrientationConceptId_ImageTable.R
\name{imageOrientationConceptId}
\alias{imageOrientationConceptId}
\title{'imageOrientationConceptId'}
\usage{
imageOrientationConceptId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing direction of each DICOM image
}
\description{
imageOrientationConceptId function will represent direction of each image
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageOrientationConceptId(DICOMList)
}
