% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageResolutionColumns_ImageTable.R
\name{imageResolutionColumns}
\alias{imageResolutionColumns}
\title{'imageResolutionColumns'}
\usage{
imageResolutionColumns(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing resolution(especially columns) of each DICOM images
}
\description{
imageResolutionColumns function will resolution(especially columns) of each images
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageResolutionColumns(DICOMList)
}
