% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageResolutionRows_ImageTable.R
\name{imageResolutionRows}
\alias{imageResolutionRows}
\title{'imageResolutionRows'}
\usage{
imageResolutionRows(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing resolution(especially rows) of each DICOM images
}
\description{
imageResolutionRows function will resolution(especially rows) of each images
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageResolutionRows(DICOMList)
}
