% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSliceThickness_ImageTable.R
\name{imageSliceThickness}
\alias{imageSliceThickness}
\title{'imageSliceThickness'}
\usage{
imageSliceThickness(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list indicating slice thickness(distance) between each images taken by CT
}
\description{
imageSliceThickness function indicates slice thickness(distance) between each images taken by CT
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageSliceThickness(DICOMList)
}
