% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTotalCount_OccurrenceTable.R
\name{imageTotalCount}
\alias{imageTotalCount}
\title{'imageTotalCount'}
\usage{
imageTotalCount(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating total count of each DICOM files in each occurrences
}
\description{
imageTotalCount function indicates total count of each DICOM files in each occurrences
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageTotalCount(DICOMList)
}
