% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modality_Occurrence_Table.R
\name{modality}
\alias{modality}
\title{'modality'}
\usage{
modality(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating modality of each shoot
}
\description{
modality function indicates modality of each shoot
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
modality(DICOMList)
}
