% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalityImage.R
\name{modalityImage}
\alias{modalityImage}
\title{'modalityImage'}
\usage{
modalityImage(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating modality of each image
}
\description{
modalityImage function indicates modality of each image
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
modalityImage(DICOMList)
}
