% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyOccurrenceId_ImageTable.R
\name{radiologyOccurrenceId}
\alias{radiologyOccurrenceId}
\title{'radiologyOccurrenceId'}
\usage{
radiologyOccurrenceId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing anonymized StudyInstanceUID of DICOM
}
\description{
radiologyOccurrenceId function will read StudyInstanceUID metadata of all DICOM files read by DICOMHeaderList function
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyOccurrenceId(DICOMList)
}
