% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyOccurrenceTable.R
\name{radiologyOccurrenceTable}
\alias{radiologyOccurrenceTable}
\title{'radiologyOccurrenceTable'}
\usage{
radiologyOccurrenceTable(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
radiologyOccurrenceTable in a dataframe format
}
\description{
radiologyOccurrenceTable function will show you radiologyOccurrenceTable in a dataframe format
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyOccurrenceTable(DICOMList)
}
