% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{RCDMShinyViewer}
\alias{RCDMShinyViewer}
\title{'RCDMShinyViewer'}
\usage{
RCDMShinyViewer(Radiology_Occurrence_Table, Radiology_Image_Table)
}
\arguments{
\item{Radiology_Occurrence_Table}{is a result of R-CDM : Radiology_Occurrence_Table<-radiologyOccurrenceTable(DICOMList)}

\item{Radiology_Image_Table}{is a result of R-CDM : Radiology_Image_Table<-radiologyImageTable(DICOMList)}
}
\value{
result of database analysis
}
\description{
RCDMShinyViewer function visualizes result of database analysis
}
\examples{
Radiology_Occurrence_Table<-radiologyOccurrenceTable(DICOMList)
Radiology_Image_Table<-radiologyImageTable(DICOMList)
RCDMShinyViewer(Radiology_Occurrence_Table, Radiology_Image_Table)
}
